[FIX/#14] 빌드 로직 에러 수정 <Suspense> 로 감싸서 에러 해결
